Moved definition of CompoundJoint class (used by testJoints.cpp and testMomentArms.cpp) to a common location.
